Change imports. Restructure module.
