+++ src/kmt/__init__.py
@@ -1,3 +1,3 @@
-from .kmt import main
+from .__main__ import main
 
 __all__ = ["main"]
