+++ src/kmt/core.py
@@ -4,12 +4,10 @@ import jinja2
 import inspect
 import copy
 import logging
-import re
 
-from . import util
-from . import yamlwrap
-
-from . import exception
+import kmt.yaml_types as yaml_types
+import kmt.util as util
+import kmt.exception as exception
 
 logger = logging.getLogger(__name__)
 
@@ -19,96 +17,6 @@ default_handlers = {}
 default_step_support_handlers = []
 default_pipeline_support_handlers = []
 default_filters = {}
-
-#
-# Manifest name lookups
-class Lookup:
-    def __init__(self, spec):
-        util.validate(isinstance(spec, dict), "Invalid specification passed to Lookup")
-
-        self.spec = spec.copy()
-
-        allowed_keys = [
-            "group",
-            "version",
-            "kind",
-            "api_version",
-            "namespace",
-            "pattern"
-        ]
-
-        errored = []
-        for key in self.spec.keys():
-            if key not in allowed_keys or not isinstance(self.spec[key], str):
-                errored.append(key)
-
-        if len(errored) > 0:
-            raise exception.PipelineRunException(f"Invalid keys or invalid key value found on lookup: {errored}")
-
-    def find_matches(self, manifests, *, current_namespace=None):
-        return self._find(manifests, multiple=True, current_namespace=current_namespace)
-
-    def find_match(self, manifests, *, current_namespace=None):
-        return self._find(manifests, multiple=False, current_namespace=current_namespace)
-
-    def _find(self, manifests, *, multiple, current_namespace=None):
-        util.validate(isinstance(manifests, list) and all(isinstance(x, Manifest) for x in manifests),
-            "Invalid manifests provided to Lookup find")
-
-        matches = []
-
-        for manifest in manifests:
-
-            info = util.extract_manifest_info(manifest)
-
-            if "group" in self.spec and self.spec["group"] != info["group"]:
-                continue
-
-            if "version" in self.spec and self.spec["version"] != info["version"]:
-                continue
-
-            if "kind" in self.spec and self.spec["kind"] != info["kind"]:
-                continue
-
-            if "api_version" in self.spec and self.spec["api_version"] != info["api_version"]:
-                continue
-
-            if "namespace" in self.spec:
-                if self.spec["namespace"] != info["namespace"]:
-                    continue
-            elif info["namespace"] is not None and info["namespace"] != current_namespace:
-                # If no namespace has been defined in the lookup, we will match on
-                # the current namespace and any resource without a namespace.
-                continue
-
-            if "pattern" in self.spec and not re.search(self.spec["pattern"], info["name"]):
-                continue
-
-            matches.append(manifest)
-
-        if multiple:
-            return matches
-
-        if len(matches) == 0:
-            raise exception.PipelineRunException("Could not find a matching object for Lookup find")
-
-        if len(matches) > 1:
-            raise exception.PipelineRunException("Could not find a single object for Lookup find. Multiple object matches")
-
-        return matches[0]
-
-
-def lookup_representer(dumper: yaml.SafeDumper, lookup: Lookup):
-    return dumper.represent_mapping("!lookup", lookup.spec)
-
-def lookup_constructor(loader: yaml.SafeLoader, node: yaml.nodes.MappingNode):
-    return Lookup(spec=loader.construct_mapping(node))
-
-yaml.SafeDumper.add_representer(Lookup, lookup_representer)
-yaml.Dumper.add_representer(Lookup, lookup_representer)
-
-yaml.SafeLoader.add_constructor("!lookup", lookup_constructor)
-yaml.Loader.add_constructor("!lookup", lookup_constructor)
 
 class Manifest:
     def __init__(self, source, *, pipeline):
@@ -122,7 +30,7 @@ class Manifest:
         self.local_vars = {}
 
     def __str__(self):
-        output = yamlwrap.dump(self.spec)
+        output = util.yaml_dump(self.spec)
 
         return output
 
@@ -385,7 +293,7 @@ class Pipeline:
         return manifests
 
     def _resolve_reference(self, current_manifest, item):
-        if isinstance(item, Lookup):
+        if isinstance(item, yaml_types.Lookup):
             current_namespace = None
             metadata = current_manifest.spec.get('metadata')
             if metadata is not None:
